Clustering is feature toggleable
Now that open365 targets multiple platforms, we need the flexibility to
perform optimally for each unique environment.  This could mean getting
more out of the CPU by launching multi-node processes in a large
environment with a lot of processing power or reducing to a single-
threaded process for smaller environments with minimal processing power.

For example, clustering should be enabled for production, but perhaps
disabled for users installing open365 in a raspberry pi.

Clustering can now be enabled/disabled with the envar
EYEOS_RABBITMQ_AUTH_BACKEND_HTTP_USE_CLUSTER